Move running servers and clients to separate classes
