+++ build.gradle.kts
@@ -15,7 +15,6 @@ repositories {
 dependencies {
     testImplementation(kotlin("test"))
     implementation("org.jetbrains.kotlinx:kotlinx-coroutines-core:$kotlinx_coroutines")
-    implementation("org.jetbrains.kotlinx:kotlinx-coroutines-debug:$kotlinx_coroutines")
 }
 
 tasks.test {
@@ -24,8 +23,4 @@ tasks.test {
 
 kotlin {
     jvmToolchain(17)
-}
-
-application {
-    mainClass.set("MainKt")
 }
