+++ src/main/kotlin/run/BlockingClientMain.kt
@@ -1,0 +1,10 @@
+package run
+
+import blocking.BlockingClient
+import utils.SERVER_ADDRESS
+import utils.SERVER_PORT
+
+suspend fun main() = runClient {
+    val client = BlockingClient(SERVER_ADDRESS, SERVER_PORT)
+    client.start()
+}
